Fix Location "Remove" button
To remove postmeta fields via XML-RPC, we need to send a custom_fields
entry with the meta ids. When a location was removed, we would just not
send the custom_fields entry, which doesn't work.

This change sets the id for the location meta fields if they exist. If
the location is being updated, we set the value. If the location is
being removed, we don't include the value. If the location is newly
being added, we set the key and value.