Enabling content selection by using a CheckableFrameLayout.
Renaming MediaSelectFragmentPagerAdapter -> MediaContentFragmentPagerAdapter.
Removing MediaContentGridView as it was an unnecessary abstraction.
ActionMode now begins when an item is long-pressed. While active the user will have the option to deselect all items, confirm the selected items to be returned to the Post editor, or create a gallery (not implemented yet).
MediaSelectActivity now returns the results of the selection fragment. The Post editor displays the results in the post.
Selecting the capture photo icon will launch the Camera to get an image taken. Not fully implemented.

Coming up:
Working image capture
Support for web images (with progress indicator and background downloading)
Video content (capture, device, and web)