Hiding Gallery menu item when no media library content is available.
To be re-enabled when image IDs can be generated when starting an upload to the media library.

This is unfortunate but is also not considered a change in behavior from current implementation. As it stands gallery creation is only possible from the Post editor with content from the users media library.