My site fragment introduced.
The top part of the fragment is completed with a few styling issues left. The site picker works with the My Site fragment now. The last used blog will be opened by default in the fragment and the user can use the switch site button and the site picker to change the blog.