Update WordPress icon to use a circle base shape.
This is based on internal discussion on the optical weight the icon carries in stock homescreens. Moving from a square to a circle allows us to stay within Google material guidelines, while still making for an icon that looks bigger next to other icons that don't necessarily follow material guidelines. It's like having the cake, and eating it too.