Moves the bucket related code to `onResume` of `NotificationsListFragment`.
This ensures that the simperium bucket will have been loaded from the `deferredInit()` method in `WordPress.java`

Also added a line in `GCMIntentService` that will run Start Simperium when a push is received. This way Simperium will still sync when a push is received (this broke with the introduction of deferredInit).

Fixes #3177