Custom notifications filter layout with wrapping Fixes #3481
This is a custom, RadioGroup-like layout that allows the filters list to wrap if space is not enough for the filter names.

Unfortunately, the filter borders are emulated (using margins and a background color for the parent container) and support for dedicated border for the selected filter is not supported. Instead, the fill-in background of the filter is made a bit more dark than it used to be. For the same reason, the rounded corners of the "radiogroup" are not supported.