Person display name styled as title
The toolbar title for the person detail page changed to "" and instead the display name styled as the title. While doing so, I've decided to get rid of the include since the changes between the list & display change is quite a bit now and we keep styling them differently. So I think it'll be easier to just have 2 different layouts that does similar things. I might look into using styling to refactor it a bit.