Allow force updating of the post list
This change allows the post list to force a UI update when it's notified
that a post has finished uploading.

- The PostsListAdapter now has a FORCED mode for loading posts, which will
update the post list UI even if the post list is apparently unchanged
- Previously, when uploading a post outside the post list (from
the editor or the preview) and returning to the list, the post would
continue to display the 'Uploading' UI even when the upload completed.
This is caused by posts no longer containing information about their
upload status internally.
- Also fixed some access modifier warnings in PostsListAdapter