Always include featured image ID when uploading a post
Simplifies the upload logic for featured images. The reason
mLastKnownRemoteFeaturedImageId and featuredImageHasChanged()
existed was metaWeblog.editPost, which threw an error when
attempting to set the featured image id to the already existing
value.

We now use wp.editPost, which doesn't have this issue, and so we
no longer have to track featured image changes, only the latest
value of the featured image id.