Add adaptive icon
This attempts to address #5529. It adds app icon back and foregrounds, and an adaptive asset.

The ongoing conversation right now is whether this looks good in the shapes other than the round version. This particular version of the icon uses a 3dp padding around the WordPress logo. Should we try a 4dp version?