Explicitly use a dummy application in Robolectric unit tests
This commit reverts to the pre-Gradle 3.0 behavior
of the Robolectric tests.

The combination of newer Robolectric + Gradle plugin 3.0 spins up the
WordPress application by default for each test, instead of using a
default 'blank' application as it did before. This broke the unit
tests, as the WordPress Application isn't set up to support this.