Remove update plugin directory list sql method
Whenever we start a new fetch, we remove the previous directories which means these models are short lived (by design). We don't need to be updating the plugin directories since there shouldn't be duplicates in the lists we get from remote. Also, if we were to update these models, it'll mess up the order and might end up confusing the user after the update. If it turns out that we get multiples for any reason, we should just ignore those results or refresh the whole table quickly to reflect all the changes in remote.