Improves the way we handle Zendesk identity in ZendeskHelper
During in one of our tests, we found that the Zendesk identity could be cleared by Zendesk SDK. This seems to happen after an identity change through another Zendesk request, for example enabling push notifications, with the old access token, which results in a 401 error. This commit makes improvements to the ZendeskHelper to handle such cases more gracefully.