Refactored PlanModel
Refactored PlanModel, by making its variables into val attributes, and initialized them through a constructor.
Updated attribute names to be more descriptive.
Changed the productId attribute as a Long instead of String.

Just to note, sometime the API return 'null' for id field. (Which is weird, since it is a long/number when it is available).