Removed Invalid token ErrorType.
API is responds with a `INVALID_TOKEN` error, if it is hit with invalid/corrupt bearer token. Also responds with `AUTHORIZATION_REQUIRED`, if the token is missing entirely.

These ErrorTypes are captured for all API as OnAuthenticationChanged event at BaseRestClient, as such these need not be included in individual API ErrorTypes.