Stop redelivering intent in the NewSiteCreationService
If the service's process is killed we don't want to restart the service
with the last intent as the service might have been in the middle of the
process and we might end up in an undefined state. For example if the
service already created a new site, restarting the service would result
in "Site already exists" api error which would crash the app.