Fixes the initial loading state of domains page
When the new domains page is visited, we try to load an initial set of suggested domains with the site's title. This search is handled differently from a regular search since we don't show this initial query in the EditText and keep showing the header until the user types a query. Also, we'll only show the clear button if the query in the EditText is non-empty.