No need to detect blocks when in Gutenberg
This way, we avoid the costly call for getting the content from
Gutenberg (which is rather slow). When we're in Gutenberg, we just offer
the switch to Aztec. When in Aztec, we only offer the GB switch if there
are blocks or the post is empty.

Props to @mzorz for the idea of optimizing the menu preparation to avoid
calling the costly Gutenberg "getContent" call.