Fixes AS warnings for getScreenTitle of new site creation fragments
AS was complaining that getScreenTitle was returning a nullable type
even though it should not. I opted to simplify the code a bit and have
the same result. I think this can be further refactored to extract at
least the throw logic somewhere, but I didn't want to make that a part
of this PR.