Update to AutoService version with separate processor
The AutoService recently split itself into two packages, with
a the annotation processor now being separate and usable via
kapt.

This addresses a Gradle 5.0 deprecation warning, but also it's
possible the explicit use of kapt + its addition to the
fluxc-processor package may address the intermittent issue of
builds failing due to unprocessed annotations.