Restore user agent for WebViewActivity
The user agent was removed in 9aa37f1 because it was causing issues like websites not showing their mobile version. Some websites use the user agent to detect whether the browser being used is from a mobile app. However, after the change in ae321eb, the `WordPress.getUserAgent()` now extends the default instead of replacing it. I believe it should be safe to use the user agent again.