Add compact information to view model
Added a view model with both after much back and forth. The problem is,
if you don't then you end up having to actually reload the data set
which is slow, because of the fact the only real cache of the data is
at the front-end (i.e. the paged adapter) so if we want to update it and
the adapter's data is changing we actually have to re-create it rather
than just re-display it. Whilst this solution is not necessarily the
most elegant, because there is information there that is not needed,
it's actually the cleanest way of doing the change without any loading /
lag.