Hide no-search-results-view when displaying themes
Previously, the theme view would display either (a) the
fetching/no-internet-connection view, or (b) both the list of themes
and (underneath that) a view saying "No themes matching
your search" (which is visible in the gaps between list items). This
fixes that by only showing the "No themes matching your search" view
when all the available themes are being filtered out by the search.