Update CONTRIBUTING.md based on the template
This uses the new `CONTRIBUTING.md` template reviewed by Editorial. Please see p2JRYi-41r-p2 for more information.

The only additional change here is the addition of the _Development Practices_ section which emphasizes the coding style and our other guides in the Wiki. 

The original content was moved to the Wiki for consistency with our other guides:

- [Using Android Resources](https://github.com/wordpress-mobile/WordPress-Android/wiki/Using-Android-Resources)
- [Pull Request Guidelines](https://github.com/wordpress-mobile/WordPress-Android/wiki/Pull-Request-Guidelines)
- [Subtree'd Library Projects](https://github.com/wordpress-mobile/WordPress-Android/wiki/Subtree'd-Library-Projects)