Temporarily adds wpColorWarningDark attr that's added in open PR: #10496
As part of merging `develop` into `master-auto-upload` branch, we need
to fix a conflict for `PostListButtonType` by changing the
`R.color.warning_dark` to an attr. This attr didn't exist yet in
`develop` but was already added in #10496. As a temporary workaround it
was suggested to add the necessary attr.