Mark PostUploadModel as CANCELLED when exiting
For self-hosted sites, when exiting the editor without uploading, the `PostUploadModel.uploadState`
can get stuck in `PENDING`. This happens in this scenario:

1. The user edits an existing post
2. Adds an image -- this creates the `PostUploadModel` as `PENDING`
3. Exits the editor by tapping on Back (not saving or publishing)

If the `uploadState` is stuck at `PENDING`, the Post List will indefinitely show a “Queued post”
label.

The `uploadState` does not get stuck on `PENDING` for WPCom because the app will automatically
start a remote auto-save when the editor exits. Hence, the `PostUploadModel` eventually gets updated.

Marking the `PostUploadModel` as `CANCELLED` when exiting should be fine for all site types since
we do not currently have any special handling for cancelled uploads. Eventually, the user will
restart them and the `uploadState` will be corrected.

See `PostListUploadStatusTracker` and `PostListItemUiStateHelper.createUploadUiState` for how
the Post List determines what label to use.