capitalizeWithLocaleWithoutLint string extension introduced
String.capitalize(Locale) call crates a lint error claiming that the default locale usage is a source of bugs. This should not happen when we are already passing a locale to it. In order to get around this issue, a wrapper extension method is introduced so we can suppress the lint warning in a single place, document the reason for it and make it easy to remove it in the future.