EditorMediaListener added.
This is an attempt to create the communicatino between the EditPostActivity and EditorMedia. Some, if not all, of the functions in EditorMediaListener are temporary, because they are not the ideal way to communicate the information. We'll need to move a lot of logic between EditPostActivity and EditorMedia to have more meaningful functions in the listener.