Resolved TalkBack bug with Notifications switch component.
The content description for the toolbar's title was being set before the toolbar's TextView was even populated so in essence, so child TextView was present. This logic was moved to the initialization function that sets the title on toolbar which creates the TextView that would further get the content description.