Only show empty view when there are no suggestions
Avoids an issue where the empty view appears when it shouldn't at times.
For example if you press the background when there is a single
suggestion and don't lift your finger, the suggestion disappears, but
the activity doesn't close until you left your finger, so the empty view
becomes visible as long as you hold your finger down.