Updates Sentry to 2.1.3.
The migration guide at https://docs.sentry.io/platforms/android/migrate/ was used to make this a 1-to-1 update. There is a very minor refactor, but otherwise the functionality should be the same.

This will currently crash because `Sentry` wants `cacheDirPath` to be passed which will be handled in a subsequent commit.