Convert type identifier override field to a one-liner
When such code fits in one line it a recommended that it become a
one-liner. Since I am seeing both occur in the codebase I decided to
add this improvement. Can revert that, by dropping the commit if it is
already decided by the team to not do have 'by lazy { ... }' statements
in one line.