Add search query field to post list descriptor for xml rpc site
As with the 'searchQuery' field on the 'PostListDescriptorForRestSite',
which is nullable and defaults to null, this newly added 'searchQuery'
is following the same structure.

This is the first feature commit, which will enable the XML RPC clients
to add a search query while making a request for a list of posts, just
like it is already enabled for the Rest clients.