Add post list descriptor test case for search query for xml rpc
This test was written on purpose before making the actual change, TDD
style, to prove that the next commit adds value. As such, the
'PostListDescriptorTest' will fail for that newly added test case at the
moment because the 'shouldHaveSameUniqueIdentifier' is actually 'true'
at the moment and not 'false' as expected.