Update search is not available for xmlrpc sites test
By enabling search for self-hosted site, this test should pass with the
current change. However, as it is now it doesn't since the main code is
not yet implemented. Thus, updating first the test to fail and in the
next commit the change will be implemented so that the test passes.