🐛 Fix stubs crashing in Arabic
{{now}} tag in stubs was using the current locale to render the date in the JSON stubs, so in non-roman languages like Arabic that generated a string that was then not parsable as a Date.

When handling dates that are meant to be locale-agnostic, like the ones we want in the fields of our JSONs, we need to use the `en_US_POSIX` language (as per ISO recommendations; that's exactly what this locale is for after all), not the locale of the user.

The fix consists of using a new tag `{{fnow}}` in our stubs, and declaring that tag as a custom extension on our WireMockRule. The implementation for this tag ensures we use `en_US_POSIX` and that we're thus locale-agnistic when generating dates in our JSON.