Move ReaderPostPagerActivity intent-filter to alias
The alias is necessary so that the system doesn't try to
destroy ReaderPostPagerActivity when we disable the deeplinking
component via PackageManager#setComponentEnabledSetting. We
instead disable the alias, which leaves the real activity alone
but disables the intent-filter deeplinking behavior as intended.

This especially matters when ReaderPostPagerActivity is the
current activity, for example when an external URL is opened
from the Reader, as this ends up closing the app.