Append timestamp to output file path
Outfile file path previously depended only on input fie path hash. If the same input file path was re-chosen for editing, already present old cached image for that file was loaded on the preview screen on returning from the crop screen. Appending timestamp to the output file path to get a unique path.