Rename prod deps gradle task to something more suited and consistent
The name 'prodDeps' for this Gradle task is a bit confusing for
developers since it might look like it can only be used for configuring
production builds. However, this task can also be used for configuring
debug builds by installing some extra dependencies. Also, the same task
on iOS is called 'configure_apply', which doesn't help with consistency.

This commit renames both this task's name and description to make sure
that it is consistent with the iOS task and also descriptive enough so
that developers can understand that this task can also be used to
configure debug builds.