Try to remove jitpack – Still fails, now it can't find Aztec either
I though maybe if I get rid of jitpack it'll use my repo for Gutenmberg
and be happy. Didn't consider the dependency on Aztec. I guess I need to
find a way to tell WordPressEditor to look for it into the Gutenmberg AAR.

Failed with

FAILURE: Build failed with an exception.

* What went wrong:
Could not determine the dependencies of task ':libs:editor:WordPressEditor:compileDebugAidl'.
> Could not resolve all task dependencies for configuration ':libs:editor:WordPressEditor:debugCompileClasspath'.
   > Could not find com.github.wordpress-mobile.WordPress-Aztec-Android:aztec:v1.3.45.
     Required by:
         project :libs:editor:WordPressEditor
   > Could not find com.github.wordpress-mobile.WordPress-Aztec-Android:wordpress-shortcodes:v1.3.45.
     Required by:
         project :libs:editor:WordPressEditor
   > Could not find com.github.wordpress-mobile.WordPress-Aztec-Android:wordpress-comments:v1.3.45.
     Required by:
         project :libs:editor:WordPressEditor
   > Could not find com.github.wordpress-mobile.gutenberg-mobiled:react-native-bridge:20201210-144401.
     Required by:
         project :libs:editor:WordPressEditor