Track content presented on reader discover screen
This solution covers both, the success and error path.

Also, this commit fixes the two our of four failed view model tests,
which was previously failing due to the missing implementation. The
extra two tests that was passing, even before, are making sure that the
content presented track event is not triggered on specific scenarios.
As such, by having them passing again, these scenarios are being
verified as well.