Feature: Update activity log page size from 10 to 20
This is done for two main reasons:
1) So that both platforms, iOS and Android, become consistent with each
other. Currently, the activity log page size on iOS is 20.
2) So that the scrolling issues on the 'Backups' screen, which depends
on the 'Activity Log' screen, become twice as hard to appear.

Note: The bullet point.2 above will not be fixing the scrolling issues
on the 'Backups' screen, just them harder to appear. To avoid any
scrolling issue and API enhancement in required. If the API enhancement
is not ready by the time this feature goes live, then an additional
increase to the page size, from 20 to 100, would be necessary. This will
make it even harder for such problems to appear until the API
enhancement becomes available.