Build: Update fluxc hash
This 'FluxC' hash updates 'FluxC' to that 'branch' version where
'ActivityLogStore.getActivityLogForSite(...)' utilises the
'rewindableOnly' to retrieve only the rewindable activities from the
database.

This step is required in order to build the 'Backups' screen. During
code review and before merging to 'develop' this hash will be replaced
with the appropriate tag.