Feature: Trigger backups screen launch
This commit adds a new launcher method, the 'viewBackupList', which
launches an 'Activity Log' screen, just like the
'viewActivityLogList(...)'. The only difference is the additional
extra boolean, the 'ACTIVITY_LOG_REWINDABLE_ONLY_KEY', which is added
prior to launching the screen and will become the differentiator between
presenting the 'Activity Log' or 'Backups' screen.