Feature: Hide activity type chip based on the rewindable only flag
If the rewindable only flag is present and true then the 'Activity Type'
chip is hidden, else it remains visible. As such, on the 'Activity Log'
screen the 'Activity Type' chip and filtering capability will be
available. But, on the 'Backups' screen the same 'Activity Type' chip
and filtering capability will not be available.

This way both screens are visually differentiated by 'Activity Type'.

Note: This commit also update the synthetic import, from 'toolbar_main'
to 'activity_log_list_activity'. It seems that the previous import was
incorrect.