Fix: Update retrieval of the rewindable only flag from activity's intent
The system recreates the activity, as such the extras will still be
available in activity's intent. As such, retrieving the 'rewindableOnly'
flag from there is a much better and robust alternative.

Note: The 'SiteModel' is a bit different since it can change. Ideally
only the 'siteId' would be stored to avoid that as well. As such, those
two cases are not comparable.