Feature: Make post restore use case types parameter nullable
This is required in order to allow running this use case from within the
activity log screen when that screen is launched with a rewind request
from the activity log detail screen, which doesn't have any 'type'
selection functionality.

See 'ActivityLogViewModel.onRewindConfirmed(...)' which only holds a
reference to the 'rewindId' and doesn't know anything about the 'types'.