Feature: Launch fetch activities rewind when restore id is null
Again, this is required in order to allow running this use case from
within the activity log screen when a web restore launch is triggered
and the user (re)launching the screen or triggered a pull-to-refresh.
When that happens the activity log screen will query for any restore
that might be happening in order to show the progress row.

However, unless the 'ActivityLogStore.fetchActivitiesRewind(...)' is
launched at least once 'ActivityLogStore.getRewindStatusForSite(...)'
will not get a valid 'rewind' object back. Thus, this fetching of the
activities rewind is necessary for that scenario to work properly.