Test: Add reload events tests for activity log view model
In order to be able to add 'reloadEvents(...)' specific tests to the
'ActivityLogViewModelTest' suite that functioned had to become
'@VisibleForTesting'. Otherwise it would have been to complicate to
create those tests.

Note that this exact need indicates that the 'reloadEvents(...)' could
be extracted to its own class in the future.