Refactor: Replace disable actions display progress with restore event
This new 'RestoreEvent' data class will hold all the necessary
information that 'reloadEvents(...)' will need to function.

This refactor actually break the functionality since 'isCompleted' is
never set to true, but tests work as expected. This is a necessary step
for further refactorings, which would ultimately help replace the
'RewindStatusService' with the 'GetRestoreStatusUseCase'.