Refactor: Replace rewinding activity with restore event date
Instead of depending on the 'RewindStatusService' to get the
'rewindingActivity', this change makes it so that the restore event will
hold the date info. Thus, this makes the whole view model less stateless
and dependant on the 'RewindStatusService'.

This refactor actually break the functionality since 'date' is
never set, but tests work as expected. This is a necessary step
for further refactorings, which would ultimately help replace the
'RewindStatusService' with the 'GetRestoreStatusUseCase'.