Refactor: Replace scoped view model with view model scope extension func
This is done for two main reasons:
1) Using Google's pure 'ViewModel' with its accompanying
'viewModelScope' KTX extension function is the recommended way, and
2) The custom 'ScopedViewModel' is using a 'Job', while 'viewModelScope'
is using a 'SupervisorJob'. The main difference between the 'Job' and
the 'SupervisorJob' is that children of a supervisor job can fail
independently of each other.

The 2nd reason above is really important for the next commit, which will
introduce another 'restore status job', in addition to the already
existing 'fetch activities job'. This new job will also be cancelled
every time before it being launched. Thus, using the current
'ScopedViewModel + Job' combination will cause the 'fetch activities
job' to cancel as well. By replacing that with 'viewModelScope +
SupervisorJob' this limitation is lifted and everything will work as
expected.