Refactor: Move flow on operator and dispatcher to backup use case
This is the most robust way of using this use case since clients do not
need to manage the dispatcher themselves. This way the use case will
always run in the background no matter which client will be using them.

Also, this change reduces some of the duplication where every client
needs to explicitly call the 'flowOn(...)' operator.