Feature: Make get backup status use case restore id parameter nullable
This is required in order to allow running this use case from within the
activity log screen when a web backup launch is triggered and the user
(re)launching the screen or triggered a pull-to-refresh. When that
happens the activity log screen will query for any backup that might be
happening in order to show the progress row.

Thus, this 'downloadId' needs to become optional for when that id is not
known in advance. And based on that change the if logic to allow
querying for status needs to be updated to allow this scenario.