Feature: Update activity log page size from 20 to 100
This is done since the API enhancement is not ready yet. As such, and
in order to avoid some of the scrolling issues on the 'Backups' screen
the activity log page size has been increase to 100 (x5 increase).

Note: This commit will not be fixing the scrolling issues on the
'Backups' screen. It will just make them harder to appear. To avoid any
scrolling issue and API enhancement in required.