Build: Integrate detekt with the wordpress android module
This commit adds 'Detekt' version '1.15.0', that bundles with Kotlin
version '1.4.10', which is the Kotlin version currently utilised within
this repo. The next version of 'Detekt' is the '1.16.0-RC1', which
updates Kotlin to '1.4.21'. But, this is unnecessary, at the moment, for
this repo. Also, it is still a release candidate.

In addition to that, this commit configures 'Detekt' using the default
configuration, defaulting all options to 'false' for ease of integrating
the tool to the repo. The 'detekt.yml' file was automatically generated
running the 'detektGenerateConfig' command. This file remained unchanged
in order to utilise 'Detekt's' default configuration as a starting
point. The only configuration that was updated was the
'warningsAsErrors', which was update to 'true' in order to have all
warning being treated as errors by default.

Apart from that, the 'detekt-formatting' plugin was also added to this
configuration in order to have the 'Formatting' section being reported,
alongside all its corresponding formatting violations (currently 45).
However, the 'autoCorrect' configuration flag was set to 'false' in
order to not have the codebase being automatically reformatted whenever
'Detekt' gets triggered.