Build: Add detekt baseline for the wordpress android module
This 'Detekt' baseline suppresses the existing issues (880 in total) to
have 'Detekt' running as if the codebase has zero issues. This helps to
start utilising 'Detekt' immediately without having to worry about the
existing issues.

Any new issues that will be added to the codebase right after will be
reported and the build will fail accordingly.