Build: Add detekt as a ci step
This 'Circle CI' step differs to the rest since it actually targets a
specific module, the 'WordPress' module, instead of running 'Detekt' on
all the modules, like the other steps do. This is by design since
'Detekt' is currently configured for the 'WordPress' module only.

Locally, every developer will need to also trigger
'./gradlew WordPress:detekt' instead of './gradlew detekt' as the
former will run 'Detekt' on all modules, which is not currently
supported.