Feature: Replace unregistered email with show signup magic link step
This change effectively removes the intermediate 'Send link by email'
screen.

This screen is actually a dual purpose screen because it is also used
for with social login (Google). As such, the screen is not removed per
se, but actually becomes now a single purpose screen to be used only for
the purpose of logging in with Google.