Refactor: Add back the got unregistered email call on login email frag.
This is done purely because otherwise the behavior for the 'Woo' app
will change, which will end up breaking the 'Woo' flows.

As such, this commit reverts that and instead updates the
'LoginActivity', which exists only on the 'WordPress' side and makes
sure that the implementing 'gotUnregisteredEmail(...)' method calls the
'showSignupMagicLink(...)' instead, thus updating only the 'WordPress'
login flow with the intended change.