Refactor: Add back the track email signup confirmation viewed method
This is done purely because otherwise this will break the 'Woo' app,
which also using this interface method, even if that is currently just a
'TODO("Not yet implemented")'.

As such, this commit reverts that and instead updates the implementing
method on the 'WordPress' side with a '\\ Do nothing.' comment.