Feature: Remove rewind status service and rewind available logic
This 'RewindStatusService' is only being used in order to determine
whether the 'activityRestoreButton' is to be shown or not. Since the
same logic needs to be applied for the 'activityDownloadBackupButton',
another, more simplified solution will substitute the service. This
other solution will pass the buttons visibility flag from the previous
screen and show/hide both the restore and download backup buttons.

Actually, after thinking on the existing service solution it might be
even better if the UX is updated to the more simplified solution since
otherwise, with the current solution, a user might see the restore
button appear out of nowhere, given the fact that they enter the details
screen while a restore is in progress, and that restore finishes while
they are within the details screen, not knowing in the end, why this
button ends up being appearing afterwards.

Obviously, the more simplified solution would not cover some edge cases.
However, those edge case, if need be, can be worked on later on.

As such, the 'RewindStatusService' is no longer needs and can be safely
removed from the codebase.