Feature: Launch reader interest screen from setting using a fab button
This solution connects the settings screen with the reader interests
screen through the newly created fab button. In addition, this changes
makes sure that when the user returns from this screen an update is
performed so that the settings screen gets reloaded with the newly
picked interests (if any). Thus, activity for result is being utilised
during the launch.