Feature: Add basic track related functions to reader tracker
These newly added basic track related functions will act as delegates
forwarding the the reader related track events to the analytics tracker.

This is done so that every reader related tracking event flows through
the 'ReaderTracker' instead of the individual analytics tracker wrapper
or utils helper classes. Going forward the 'ReaderTracker' class will be
responsible for all the tracking that is happening within the 'Reader'
screens and features.