Refactor: Replace track with site id with track blog id
This actually changes functionality a bit since 'trackWithSiteId(...)'
is checking whether the 'blogId' is different to zero (0) and if so it
doesn't add this property. However, the 'track.BlogId(...)' function
will always add the property no matter if it is zero (0) or not.

This has been discussed and agreed as there is no point in excluding the
zero (0) blog id properties from being tracked. As such the code gets
simplified.