Refactor: Replace semantic with primary colors on wordpress base colors
This way the primary and secondary base colors end up being defying in a
more robust way since their definition is not being duplicated but
actually reused through the primary colors. This effectively means that
if a new requirement comes in to change the primary 'blue_50' to, for
example, a 'red_50', then this change will only happen on the
'primary_50' attribute and then automatically reused by the primary and
secondary base colors. Else, the 'red_50' change would need to have been
applied to primary and secondary base colors explicitly. Thus, making
the colors structure more intuitive and easier to reason about.