Feature: Remove dark theme version of the wordpress tab layout style
This dark theme version of the 'WordPress.TabLayout' style is actually
overriding the light theme one, which causes the tab bar night colours
to default to the original ones. This is because they are not picking up
the 'tabTextColor', 'tabSelectedTextColor' and 'tabIndicatorColor'
values as it seems was expected.

Apart from that, the dark vs. the light theme are identical, with the
'android:layout_marginBottom' configuration being the exception.
However, this extra configuration is confusing and needs revisiting.

By removing the dark theme version of the 'WordPress.TabLayout' style
altogether, the problem is fixed automattically. For the 'WordPress'
dimension, in the dark theme, the text is no longer 'blue', instead it
is 'white', as expected, with the line remaining as 'blue'. Same for the
'Jetpack' dimension, the text ends up being 'white' and the line remains
'green'.