Fix: Replace instance with static reader tracker
Since the 'trackTagLoaded(...)' method is also being triggered through
the static 'newInstanceForTag(...) method, this means that the
'mReaderTracker' field instance will not be yet available, and as thus
cannot be used, or else it will result in a null pointer exception.

This commit replaces the instance with the static reader tracker
equivalent function to mitigate this situation and resolve any potential
crash.