Feature: Remove mipmap m dpi beta app icon
This is done in order to match the WordPress 'wasabi' variant, which
doesn't have a 'mipmap-mdpi/app_icon.png' file as well.

Actually, all 'mdpi' and 'ldpi' resources are no longer being utilises
as much with the latest Android devices and as such they are becoming
more and more obsolete these days.