Feature: Remove mipmap m dpi alpha app icon
This is done for the WordPress 'jalapeno' variant as well, since it was
probably omitted in the first place, just like it was done on the rest
of the `WordPress` variants.

Actually, all 'mdpi' and 'ldpi' resources are no longer being utilises
as much with the latest Android devices and as such they are becoming
more and more obsolete these days.