Feature: Remove the extra switch blue medium theme and its usages
This seems to be the only exception to the 'colorControlActivated' rule
that apply to every ui component within the codebase (like checkboxes,
switches, etc). As such, and in order to keep consistency across all the
components, without any exceptions whatsoever, this 'SwitchBlueMedium'
theme is now removed.

Note: This 'SwitchBlueMedium' exception was added as part of this PR:
- PR: Issue/update switch color #7857
- Link: https://github.com/wordpress-mobile/WordPress-Android/pull/7857