Add hasSite check before sites deletion on sign out
Sign out is triggered on launch of the no sites error screen. On sign out, the sites are removed from the site store and onSiteRemoved event is triggered. As part of the site removed handling, when no sites are available, sign in screen is redisplayed. This leads to flashing of the no sites error screen followed by the sign in screen.

To prevent this flashing, hasSite check is added before sites deletion on sign out so that sites are removed only when they're available.