Refactor if/else statement in onSetFeaturedImageButtonClicked()
With this commit, the if/else statement is refactored to a series of single if statements, with accompanying comments. The goal is to make the code more readable and for it to be clearer what the purpose of each combination is. It also adds an explicit if statement for any cases where the mediaId may equal mFeaturedImageId.