Introduce "MEDIA_ID_NO_FEATURED_IMAGE_SET" const
With this commit, a "MEDIA_ID_NO_FEATURED_IMAGE_SET" constant is introduced and used in checks where a featured image is being removed or is already set to zero.

The purpose behind introducing a constant is to make its purpose clearer and make it easier to change in the future, if needed.