Remove editor check from UBE
The `/sites/{site_id}/gutenberg` endpoint is unreliable (see https://github.com/wordpress-mobile/gutenberg-mobile/issues/3425#issuecomment-840223555) but was being used to disable the UBE on sites where Gutenberg was unavailable.

Since most users will have Gutenberg enabled in this scenario, it makes sense to remove usage of this API endpoint and just show a failure state if Gutenberg is unavailable. The failure state will show information on how to fix the error (i.e. enable Gutenberg on the site).