Update Tracks event to represent exit
When the suggestions UI is exited, a value of "false" should be sent back with the tracks event. See: p9ugOq-1tR-p2

With this commit, an if/else statement is added to "finishWithValue()" to send that value of "false" in cases where only a space has been passed to the function (indicating the user has simply existed the UI).