Analysis: Remove 1 vector path issue from baseline
There are already 119 'VectorPath' performance related warning in the
lint report list. As such, having this 1 extra vector path issue in the
baseline is not necessary. Also, this way, all vector path issues will
become available in the lint report list, in a single place, which will
become the single source of truth for that matter.