Refactor 'id' param so it isn't bound to a type
The 'id' param is currently bound to an Int type. This reflects the fact that it's currently only being used to generate one dialog, which passes a mediaId. Ideally, it should be possible to pass any piece of data to the class from Gutenberg. See this comment for more context: https://github.com/wordpress-mobile/WordPress-Android/issues/14676#issuecomment-852024183

With the changes proposed in this commit, the 'id' param is refactored to make use of generics, so that it isn't bound to any specific type.

It is also renamed to 'dataFromGutenberg', in an attempt to make its purpose clearer, which is to pass in any piece of data from the editor to the dialog.