Analysis: Remove lint baseline
The only active issues within the Lint baseline are those 4/5
'ClickableViewAccessibility' related ones, which are also very specific
to the 'AztecEditorFragment.java' class, and those can be suppressed
closer to the source (for now).

As such, the baseline can be now safely removed as the remaining issues
will be also suppressed in a subsequent commit.