Analysis: Suppress overdraw lint issues for jetpack dimension
These extra 'Overdraw' Lint issues is now suppressed so that the
'wordpress' and 'jetpack' dimension becomes equivalent, in terms of the
total and individual number of Lint warnings.

PS: Actually all the 'Overdraw' Lint issues should be suppressed and
be dealt with via another path. Then, when there will no longer be any
Lint warnings available, and in order to stop adding new Lint warning
to the codebase, the 'warningsAsErrors' Lint configuration can be turned
on. This will make sure that, just like with new errors, any new warning
that is introduced should be immediately treated, either fixed or
suppressed.