gradle: the detekt configuration is now applied by the plugin a8c-apply-style-gradle-plugin
This is a proof of concept that shows how we can apply a plugin with custom
common defaults and configuration files into a project.

At first we're applying the detekt plugin and the configuration is taken
care by new plugin a8c-apply-style-gradle-plugin [1]

At first you should clone the plugin's code as the plugin is not published
anywhere:

  git clone git@github.com:potyl/apply-style-gradle-plugin.git

It is recommended to clone the repo as a sibling of this project.

You can then apply the plugin with:

  ./gradlew --include-build ../apply-style-gradle-plugin WordPress:detekt

[1] https://github.com/potyl/apply-style-gradle-plugin