Feature: Remove unnecessary home as up indicator and subtitle properties
This 'home indicator' change was done because changing the default back
icon with the close icon is actually not necessary. This is because
every other screen is using the same back icon instead of this close
icon, as such this change increases the consistency between the screens.
For example, the 'Site Settings' screen is using the back icon as well.

Also, the 'subtitle' change was done because setting the subtitle to
empty is not necessary as well. The diff is absolutely the same.