Add Jetpack check to unsupportedBlockEditorSwitch
If the 'unsupportedBlockEditorSwitch' boolean evaluates to 'true', it will display an option to sign in via SSO when attempting to access the UBE.

As the boolean was only checking for whether SSO is already enabled, it was evaluating to 'true' for self-hosted sites that didn't have Jetpack. This caused the issue outlined here: https://github.com/WordPress/gutenberg/pull/32395#pullrequestreview-694360310

With this commit, an extra check is added to ensure this boolean only evaluates to 'true' when a site is actually a Jetpack site.