Refactor logic around "space handling"
This commit refactors the logic that handles how a space is handled in the suggestions UI, as well as how that space is tracked.

The reason for this refactor is that the logic was previously split up into two different places, making it potentially tricky for future devs to parse. With this commit, it's handled in the "TextWatcher", with a boolean being passed to the "finishWithValue" function.