Refactor: Init site picker adapter during on create
Instead of initializing the 'SitePickerAdapter' during
'setupContent(...)', this is now done during the 'onCreate(...)' phase
of this screen's lifecycle. This is due to the fact that this adapter
and its publicly made 'getBlogsForCurrentView()' method will be required
to be available during 'createMainView(...)' so that the appropriate new
login epiloge screen is being picked up (the normal or expanded one).