Feature: Add is quick start done util wrapper function
This newly added function, in addition to the already existing
'isQuickStartInProgress(...)' fucntion, will help decide whether the
quick start prompt is to be shown when a site is being selected from
within the login epilogue screen.

The 'isQuickStartInProgress(...)' will help identify whether the quick
start is in progress, and if so, the quick start prompt will not show.

The 'isQuickStartDone(...)' will help identify whether the quick start
is done, and if so, the quick start prompt will not show.

As such, the quick start prompt will only shown when the quick start is
not in progress and it is not already done, thus achieving the decired
logic of showing the quick start to already existing users or new ones.