Refactor: Use site picker activity key local id key on login epilogue
This is done in order to make sure that the same
'SitePickerActivity.KEY_LOCAL_ID' key is used both during the write and
read operation on that specific flow.

Since the read happens during 'WPMainActivity.setSite(...)' and the key
that is used there is the 'SitePickerActivity.KEY_LOCAL_ID' then it
makes sense to use the same key on the 'LoginEpilogueActivity' as well
and avoid duplication, which might cause sync problems afterwards.