Feature: Do not skip quick start when no thanks option is chosen
After some thorough conversation is has been decided to always show the
quick start dialog, when coming from the login epilogue screen. As such,
clicking on the 'No thanks' option doesn't need to alter any quick start
state and can default to only triggering a track event, exactly as it
used to behave.

This is done because the current solution anyways doesn't support saving
the quick start state during re-login, which would enable knowing
whether the quick start is in progress/done for a specific site.

The current solution is based on the 'Local Site Id', which changes for
a specific site after a re-login. As such, any quick start state that
was based on that is being unmapped during re-login and then stays
unmapped for ever.

In order to overcome this problem, the current solution needs to be
migrated to another solution, one that is based on the '(Remote) Site
Id' and 'Self Hosted Site Id' instead. However, this is a major task by
itself and as such dealt with accordingly if need be.