Fix: Set site data and show quick start dialog for result ok only
If the 'setSite(...)' and 'showQuickStartDialog()' functionality is not
wrapped around checking if 'RESULT_OK' is issued then weird bugs can
appear like showing setting the site and showing the quick start prompt
even when the user clicks back or swipes to dismiss the epilgue screen.
This simple fix prevents those weird app states from showing when user
is trying to navigate throught the app.