Update related posts and post actions only when the selected post is displayed
Related posts and post actions are requested when the selected post is already fetched and laid on the screen. At this time, it is expected that the ui state is in ReaderPostDetailsUiState.

If, for any reason, the ui state is not in ReaderPostDetailsUiState, it is an illegal state and the related posts and post actions should not be updated.