Feature: Show existing epilogue screen for jetpack app
The new epilogue screen(s) is only to be showed for the WordPress app.
Since on the Jetpack app it has been decided to not allow users to
create new site (at least for now), it has been also decided that the
new epilogue screen will not be shown. This is done because the new
epilogue screen also allows for new site creation.

As such, for the Jetpack app, the quick start prompt is never shown as
well. But, since this fullscreen dialog is now shown at all, any logic
related to that and excluding it for the Jetpack app is ignored as
unreached.